fix: crash-hard utilization, int64 attention/mamba3 gfx942, stale tpu_v5p in C++
utilization.py: replace all graceful degradation with hard crashes — missing
device/sampler/workload raises RuntimeError with attributable message. No
exception swallowing, no null-metric fallbacks, no status-flag degradation.

gfx942 attention: widen 4 base-offset calculations and 4 LSE index
expressions from int32 to int64_t (matching sm_90 pattern) to prevent
silent OOB on >2^31-element attention tensors.

gfx942 mamba3: widen 2 base-offset calculations from int32 to int64_t.

C++ stale tpu_v5p: update fused_wired_cells.inc (33 TPU cell strings),
dispatch.cpp arch fallback, and bindings.cpp docstring from tpu_v5p to
tpu_v6e.

Self-tests: replace test_track_cell_degrades_gracefully with
test_track_cell_crashes_on_no_device (asserts RuntimeError). Doc counts
reconciled to 156 across BUILD_REPORT, PHASE7_REPORT, HARDWARE_VALIDATION,
verify_stage0.sh.

Verified: self-test 156/0, ruff clean, drift guard PASS, AMDGCN 17/17.

+++ grokking_optimizers/kernels/gfx942/mamba3_gfx942.hip.hpp
@@ -1092,7 +1092,7 @@ void mamba3_conv1d_forward(
         float w2 = to_float(conv_w[ch * 3 + 2]);
         float bias = to_float(conv_b[ch]);
 
-        int base = b * SEQ_LEN * d_inner + ch;
+        int64_t base = static_cast<int64_t>(b) * SEQ_LEN * d_inner + ch;
 
         float x_left  = (t > 0)           ? to_float(x_main[base + (t - 1) * d_inner]) : 0.0f;
         float x_center =                    to_float(x_main[base + t       * d_inner]);
@@ -1802,7 +1802,7 @@ void mamba3_conv1d_backward(
         float w2 = to_float(conv_w[ch * 3 + 2]);
 
         // SiLU backward: d(SiLU(u))/du = sig(u) + u*sig(u)*(1-sig(u))
-        int base = b * SEQ_LEN * d_inner + ch;
+        int64_t base = static_cast<int64_t>(b) * SEQ_LEN * d_inner + ch;
         float x_left   = (t > 0)           ? to_float(x_in[base + (t-1)*d_inner]) : 0.0f;
         float x_center =                     to_float(x_in[base + t    *d_inner]);
         float x_right  = (t < SEQ_LEN - 1) ? to_float(x_in[base + (t+1)*d_inner]) : 0.0f;
